refactor vue
