+++ src/layout/theme.liquid
@@ -50,7 +50,6 @@
 
   <a class="in-page-link visually-hidden skip-link" href="#MainContent">{{ 'general.accessibility.skip_to_content' | t }}</a>
   <div id="app">
-    <div class="abs x y fill z9" v-if="visibleModal" @click="dismissModal(visibleModal)"></div>
     {% section 'header' %}
 
     <main role="main" id="MainContent" class="body-wrapper min-full-height">
