+++ src/scripts/core/is-touch.js
