+++ src/scripts/core/modal-controller.js
