--- src/scripts/templates/index.js
@@ -1,4 +1,0 @@
-import {load} from '@shopify/theme-sections';
-import '../sections/product';
-
-load('*');
