+++ src/snippets/product-grid.liquid
@@ -1,6 +1,6 @@
-<div class="collection__grid">
+<div class="collection__grid" data-vue-instance>
   {% for product in products %}
-    <product-card :product='{{ product | json }}' :link='{{ product.url | within: collection | json }}' :options='{{ product.options_with_values | json }}'></product-card>
+    <product-card :cart="cart" :product='{{ product | json }}' :link='{{ product.url | within: collection | json }}' :options='{{ product.options_with_values | json }}'></product-card>
   {% else %}
     <p>no matches</p>
   {% endfor %}
