+++ src/snippets/side-cart.liquid
@@ -1,6 +1,37 @@
-<transition v-cloak name="slide-left">
-  <div v-show="visibleModal === 'cart'" class="side-cart z10 right top bottom y fix">
-    <cart :header="true"></cart>
-  </div>
-</transition>
+<div v-cloak class="cart bg-white f fdc y">
+  {% if header == true %}
+    <div class="cart__header f aic jcb f-0">
+      <h2 class="uppercase">Cart</h2>
+      <button class="cursor-pointer cart__close" type="button" v-on:click="modalController.dismissModal('cart')">
+        <x-button></x-button>
+      </button>
+    </div>
+  {% endif %}
 
+  <div class="cart__inner f-1">
+    <div v-if="cart.error">
+      <p class="mb1">There was a problem! <span class="mb1" v-text="cart.error"></span></p>
+    </div>
+    <div v-if="cart.state && cart.state.items && cart.state.items.length > 0" class="cart__grid">
+      <cart-item
+        :cart="cart"
+        v-for="(item, index) in cart.state.items"
+        :item="item"
+        :index="index"
+        :key="item.id"></cart-item>
+    </div>
+
+    <div v-else>
+      Your cart is empty.
+    </div>
+  </div>
+
+  <div class="cart__bottom x right left bottom z1 bg-white">
+    <div class="cart__bottom__grid">
+      <div class="cart__bottom__inner">
+        <div class="f aic jcb"><span>Subtotal</span><span>$100.00</span></div>
+        <button class="button button--full">Checkout</button>
+      </div>
+    </div>
+  </div>
+</div>
