+++ src/templates/cart.liquid
@@ -1,3 +1,3 @@
-<div class="cart-page full-height">
-  <cart :header="false"></cart>
+<div class="cart-page full-height" data-vue-instance>
+  <cart :cart="cart" :modal-controller="modalController" :header=false></cart>
 </div>
